Add possibility to retrieve undefined enum value
Reference to Microsoft.CSharp NuGet is added because the code
uses "dynamic" keyword and without the reference, the code won't
compile.

Using "dynamic" keyword significantly reduces required lines of code,
but the code remains clear enough. Thanks to the unit tests, it's
verified that the "dynamic" keyword works as expected.